requirements.txt
# AI Avatar Backend Dependencies

# Web Framework
Flask==3.0.0
flask-cors==4.0.0
flask-sock==0.7.0

# AI & ML
openai==1.6.1
python-dotenv==1.0.0

# WebSocket
simple-websocket==1.0.0

# HTTP Requests
requests==2.31.0
aiohttp==3.9.1

# Audio Processing
pydub==0.25.1
soundfile==0.12.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Development
pytest==7.4.3
black==23.12.1
flake8==7.0.0
